win: fix compilation warning on MinGW
src/win/util.c: In function 'uv_os_homedir':
src/win/util.c:1189:46: warning: passing argument 3 of 'GetUserProfileDirectoryW' from incompatible pointer type
if (!GetUserProfileDirectoryW(token, path, &bufsize)) {
^
In file included from src/win/util.c:39:0:
C:/buildtools/vendor/mingw64/x86_64-w64-mingw32/include/userenv.h:52:29: note: expected 'LPDWORD' but argument is of type 'size_t *'
USERENVAPI WINBOOL WINAPI GetUserProfileDirectoryW(HANDLE hToken,LPWSTR lpProfileDir,LPDWORD lpcchSize);
^

PR-URL: https://github.com/libuv/libuv/pull/388

+++ src/win/util.c
@@ -1159,7 +1159,7 @@ int uv_getrusage(uv_rusage_t *uv_rusage) {
 int uv_os_homedir(char* buffer, size_t* size) {
   HANDLE token;
   wchar_t path[MAX_PATH];
-  size_t bufsize;
+  DWORD bufsize;
   size_t len;
   int r;
